Add files via upload
This appendix provides additional details over the verifiable conditions used to guarantee asymptotic stability in Caputo fractional-order systems. It provides further explanations to justify the selection of the least sufficient condition within the proposed neural network-assisted stability analysis method, such that it is able to find a valid $W$ which allows the least sufficient condition to enable a broader range of initial states under which the system can converge.